Изменена структура БД.
Удалены Faculty, Department, Direction
Группа теперь принадлежит только университету
Изменена структура прав
Material, Group, University, Event теперь cвязываются с Right. То есть каждая сущность теперь не имеет собственную таблицу прав.
Campus, Subject, Teacher, ClassTime - теперь связаны только c Timetable. Доступ к ним определяется по правам доступа к Timetable.
Маршруты по получению Campus, Subject, Teacher, ClassTime теперь выглядят так: /timetables/:timetableId/entities/entityId
Доступ к профилю пользователя теперь доступен по пути /users/:userId
Расписания (Timetables), которые относятся к роли "All" могут получить неавторизованные пользователи

+++ controllers/TokenController.js
@@ -32,8 +32,24 @@ class TokenController {
           return res.sendStatus(401)
         }
 
-        const newAccessToken = generateAccessToken({user: data.user})
-        const newRefreshToken = generateRefreshToken({user: data.user})
+        const user = await db.User.findByPk(data.user.id)
+
+        if (!user) {
+          return res.sendStatus(401)
+        }
+
+        const payload = {
+          user: {
+            id: user.id,
+            firstName: user.firstName,
+            lastName: user.lastName,
+            groupId: user.groupId,
+            universityId: user.universityId
+          }
+        }
+
+        const newAccessToken = generateAccessToken(payload)
+        const newRefreshToken = generateRefreshToken(payload)
 
         await db.Token.destroy({where: {userId: data.user.id}})
         await db.Token.create({token: newRefreshToken, userId: data.user.id})
